+++ src/entity/Woodcutter.cpp
@@ -176,7 +176,7 @@ namespace Game3 {
 		harvestingTime = 0.f;
 	}
 
-	void Woodcutter::harvest(float delta) {
+	void Woodcutter::harvest(double delta) {
 		if (HARVESTING_TIME <= harvestingTime) {
 			harvestingTime = 0.f;
 			auto realm = getRealm();
